+++ h3ron-h3-sys/src/lib.rs
@@ -1,7 +1,6 @@
-//! low-level bindings to H3
+//! Low-level bindings to H3
 //!
-//! This crate includes the C sources for libh3, so compiling it requires a C toolchain and the `cmake`
-//! build tool.
+//! This crate includes the C sources for libh3, so compiling it requires a C toolchain.
 //!
 #![allow(non_upper_case_globals)]
 #![allow(non_camel_case_types)]
@@ -14,4 +13,8 @@
 #[cfg(feature = "geo-types")]
 pub mod geo;
 
+#[cfg(feature = "bindgen")]
 include!(concat!(env!("OUT_DIR"), "/bindings.rs"));
+
+#[cfg(not(feature = "bindgen"))]
+include!("prebuild_bindings.rs");